requirements.txt
numpy
tensorflow
opencv-python
pillow
matplotlib
pandas
keras
jupyter
openai
flask
weaviate-client
python-dotenv
tqdm
flask-cors
